refactor #34 : 컨벤션에 따른 파일 및 폴더명 케밥케이스로 변경
