+++ src/app.tsx
@@ -1,6 +1,6 @@
 import { Global } from '@emotion/react';
 
-import FeedbackResultPage from './features/feedback-result/components/FeedbackResult';
+import FeedbackResultPage from './features/feedback-result/components/feedback-result';
 import { globalStyles } from './styles/global-styles';
 
 function App() {
