+++ src/common/components/progress-bar/progress-bar.tsx
