+++ src/features/sidebar/components/menu-buttons/menu-trigger-button.styles.ts
@@ -1,0 +1,6 @@
+import { css } from '@emotion/react';
+
+export const sidebarToggleButton = css`
+  background-color: white;
+  border: none;
+`;
