+++ src/features/sidebar/components/menu-buttons/single-menu-button.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const singleMenuButton = css`
+  padding-left: 1rem;
+  border: none;
+  background-color: transparent;
+`;
