+++ src/features/sidebar/utils/adapt-accordion-format.ts
@@ -1,4 +1,4 @@
-import { SidebarListType } from '../types/sidebarInfoTypes';
+import { SidebarListType } from '../types/sidebar-Info-types';
 
 /* 도메인 데이터를 AccordionMenu 포맷으로 변환하는 어댑터 함수 **/
 export function adaptToAccordionFormat(sidebarList: SidebarListType) {
